+++ src/pages/Router.js
@@ -2,8 +2,29 @@ import { BrowserRouter, Routes, Route, Link } from "react-router-dom"
 import  Home  from "./Home"
 import  Profile  from "./Profile"
 import Chats from "./Chats"
+import { useState } from "react"
+import { AUTHOR } from "../Components/common"
+
+const chatList = {
+    id1: {
+        name: 'chat1',
+        messages: [{ text: 'Massage first', author: AUTHOR.bot}]
+    },
+    id2: {
+        name: 'chat2',
+        messages: [{ text: 'Massage second', author: AUTHOR.me}]
+    },
+    id3: {
+        name: 'chat3',
+        messages: [{ text: 'Massage third', author: AUTHOR.bot}]
+    }
+}
+
+
 
 const Router = () => {
+
+    const [chats, setChats] = useState(chatList)
     return (
     <BrowserRouter>
 
@@ -29,7 +50,8 @@ const Router = () => {
     <Routes>
         <Route path="/home" element={<Home />} /> 
         <Route path="/profile" element={<Profile />} />
-        <Route path="/chats/:chatId" element={<Chats />} />
+        <Route path="/chats/:chatId" element={<Chats chats={chats} />} />
+        <Route path="*" element={<Chats chats={chats} />} />
 
     </Routes>
